test: refactor test files to use parameterization
