+++ tests/test_generation.py
@@ -1,80 +1,14 @@
-from pathlib import Path
+import pytest
 
-import tomli
-
-
-def get_all_files(path: "Path | str", root: Path = None):
-    if isinstance(path, str):
-        path = Path(path)
-    if root is None:
-        childs = list(path.iterdir())
-        assert len(childs) == 1
-        root = childs[0]
-    for p in path.iterdir():
-        if p.is_file():
-            if p.name == "pyproject.toml":
-                validate_pyproject_toml_generated(p)
-            yield str(p.relative_to(root))
-        else:
-            yield from get_all_files(p, root)
+from .data import SETUP_TOOLS_FILES, POETRY_FILES
+from .helpers import get_all_files
 
 
-def validate_pyproject_toml_generated(path: Path):
-    from validate_pyproject import api
-
-    # let's assume that you have access to a `loads` function
-    # responsible for parsing a string representing the TOML file
-    # (you can check the `toml` or `tomli` projects for that)
-    with path.open() as fr:
-        pyproject_as_dict = tomli.loads(fr.read())
-
-    # now we can use validate-pyproject
-    validator = api.Validator()
-    validator(pyproject_as_dict)
-
-
-def test_pip(bake_cookie):
-    out = bake_cookie(package_manager="pip")
+@pytest.mark.parametrize("package_manager,files_expected", [
+    ["setuptools", SETUP_TOOLS_FILES],
+    ["poetry", POETRY_FILES]
+])
+def test_pip(bake_cookie, package_manager, files_expected):
+    out = bake_cookie(package_manager=package_manager)
     files = set(get_all_files(out.dst_path))
-    assert files == {
-        '.copier-answers.yml',
-        '.editorconfig',
-        '.gitattributes',
-        '.github/issue_template.md',
-        '.github/pull_request_template.md',
-        '.github/release-drafter.yml',
-        '.github/workflows/release.yml',
-        '.github/workflows/test.yml',
-        '.gitignore',
-        '.pre-commit-config.yaml',
-        'LICENSE',
-        'README.md',
-        'pyproject.toml',
-        'setup.py',
-        'tests/__init__.py',
-        'tests/test_xontrib.py',
-        'xontrib/my_prompt.py'
-    }
-
-
-def test_poetry(bake_cookie):
-    out = bake_cookie(package_manager="poetry")
-    files = set(get_all_files(out.dst_path))
-    assert files == {
-        '.copier-answers.yml',
-        '.editorconfig',
-        '.gitattributes',
-        '.github/issue_template.md',
-        '.github/pull_request_template.md',
-        '.github/release-drafter.yml',
-        '.github/workflows/release.yml',
-        '.github/workflows/test.yml',
-        '.gitignore',
-        '.pre-commit-config.yaml',
-        'LICENSE',
-        'README.md',
-        'pyproject.toml',
-        'tests/__init__.py',
-        'tests/test_xontrib.py',
-        'xontrib/my_prompt.py'
-    }
+    assert files == files_expected
